requirements.txt
# Core dependencies
openai
python-dotenv
tiktoken
pandas

# Required for Word document conversion
pandoc>=0.0.1

# Testing dependencies
pytest